Create groups.py
A lot of the items in the game can be grouped together.  Right now I only have the ghosts grouped together in this file.  I should probably just put this in the ghosts.py file instead.  I also have the pellets grouped together, but that is in the pellets.py file.