Create nodemovement3.py
This is the way entities travel from node to node smoothly.  They will travel past a node if there is a neighbor in that direction and no other valid keys are being pressed.  Inherits from the abstract node movement which just has common methods that all forms of node to node movement has.